Added Buttons Functionalities
Updated Quiz Page after answering a question, if correct it goes to next question, if wrong it shows a pop up to tell that he is wrong. Counter is included to keep track of number of correct answers.
If a user skips a questions then it just go to next question.