Fixed Login and starting quiz bugs
Fixed Login bug where it was not saving the token in the shared preferences but now it does.

Starting quiz after restarting was causing it to end immediately and quiz is considered finished and data is sent with POST request, it was modified by creating async function to return future and call .then on the rest.

Edited a requirement where Wrong answer should restart the quiz not resume after trying again and gave the user the choice to restart or to go to the home page.